start update of application modules
* move dv/v functions out of noise_module.py into dvv_utils.py
* update script measure_dvv, but this should be changed into a notebook.
